pyproject.toml
[project]
name = "english-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "seblful", email = "alesha.chimba@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "mistralai>=1.2.0",
    "pillow>=12.1.0",
    "tqdm>=4.66.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "typer>=0.21.1",
    "opencv-python>=4.13.0.92",
    "python-telegram-bot>=21.0",
    "langchain>=0.3.0",
    "langchain-qwq>=0.1.0",
    "langsmith>=0.1.0",
    "jinja2>=3.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
